ZCS-2572:iOS CardDAV addressbook for Mountpoint
Was only treating mountpoint addressbooks as remote collections.
Now treat them also as addressbooks.  With this change, iOS 11 Contacts
application is able to see a shared addressbook, create a contact in it
and modify it.
Note that High Sierra Contacts can't see shared addressbooks - this is
because only one addressbook is allowed.

AddressBookCollection:

* Moved fields to top
* sharable code moved to static method

RemoteAddressbookCollection

* New subclass of RemoteCollection
* combines features of RemoteCollection and AddressBookCollection

UrlNamespace - Call the new constructor where appropriate