ZCS-2572:TestCalDav Better DebugConfig support
Support working with:
   DebugConfig.enableDAVclientCanChooseResourceBaseName set
Used this for reducing noise of failures with caldavtester.
It isn't suitable for production although we should implement
support for the client being able to PUT to its own choice of
URL for better IOP at some point.

Since this tweak was added, we escape @ in URLs more than we used to and
the tests didn't reflect that.